graph-einstellungen als page unter einstellungen per mqtt
