+++ web/themes/dark19_01/livechart.js
@@ -39,6 +39,7 @@ var all6p;
 var all7p;
 var all8p;
 var hidehaus;
+var myLine;
 
 function loadgraph() {
 	var lineChartData = {
@@ -62,7 +63,7 @@ function loadgraph() {
 			data: alp2,
 			yAxisID: 'y-axis-1'
 		} , {
-			label: 'Bezug',
+			label: 'EVU',
 			borderColor: "rgba(255, 0, 0, 0.7)",
 			backgroundColor: "rgba(255, 10, 13, 0.3)",
 			borderWidth: 1,
@@ -80,7 +81,7 @@ function loadgraph() {
 			data: apv,
 			yAxisID: 'y-axis-1'
 		}  , {
-			label: 'Speicherleistung',
+			label: 'Speicher',
 			borderColor: 'orange',
 			backgroundColor: "rgba(200, 255, 13, 0.3)",
 			fill: true,
@@ -391,6 +392,54 @@ function putgraphtogether() {
 	}
 }  // end putgraphtogether
 
+function updateGraph(dataset) {
+	var lines = dataset.split("\n");
+	for (var i = 0; i < lines.length; i++) {
+		var ldate = lines[i].split(",")[0];
+		var lbezug = lines[i].split(",")[1];
+		var lpv = lines[i].split(",")[3];
+		var llp2 = lines[i].split(",")[5];
+		var lspeicherl = lines[i].split(",")[7];
+		var lsoc = lines[i].split(",")[9];
+		var lspeichersoc = lines[i].split(",")[8];
+		var lpa = lines[i].split(",")[2];
+		var llp1 = lines[i].split(",")[4];
+		var lsoc1 = lines[i].split(",")[10];
+		var lhausverbrauch = lines[i].split(",")[11];
+		var lverbraucher1 = lines[i].split(",")[12];
+		var lverbraucher2 = lines[i].split(",")[13];
+		var lp3 = lines[i].split(",")[14];
+		var lp4 = lines[i].split(",")[15];
+		var lp5 = lines[i].split(",")[16];
+		var lp6 = lines[i].split(",")[17];
+		var lp7 = lines[i].split(",")[18];
+		var lp8 = lines[i].split(",")[19];
+	}
+	myLine.data.labels.push(ldate.substring(0, ldate.length -3));
+	myLine.data.datasets[2].data.push(lbezug / 1000);
+	myLine.data.datasets[3].data.push(lpv / 1000);
+	myLine.data.datasets[4].data.push(lspeicherl / 1000);
+	myLine.data.datasets[5].data.push(lspeichersoc);
+	myLine.data.datasets[6].data.push(lsoc);
+	myLine.data.datasets[0].data.push(llp1 / 1000);
+	myLine.data.datasets[1].data.push(llp2 / 1000);
+	myLine.data.datasets[7].data.push(lsoc1);
+	myLine.data.datasets[8].data.push(lhausverbrauch / 1000);
+	myLine.data.datasets[9].data.push(lverbraucher1 / 1000);
+	myLine.data.datasets[10].data.push(lverbraucher2 / 1000);
+	myLine.data.datasets[11].data.push(lpa / 1000);
+	myLine.data.datasets[12].data.push(lp3 / 1000);
+	myLine.data.datasets[13].data.push(lp4 / 1000);
+	myLine.data.datasets[14].data.push(lp5 / 1000);
+	myLine.data.datasets[15].data.push(lp6 / 1000);
+	myLine.data.datasets[16].data.push(lp7 / 1000);
+	myLine.data.datasets[17].data.push(lp8 / 1000);
+	myLine.data.labels.splice(0, 1);
+	myLine.data.datasets.forEach(function(dataset) {
+		dataset.data.splice(0, 1);
+	});
+	myLine.update();
+}
 
 $(window).focus(function() {
     // if the browser window gets focus again after being blurred,
